requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.7
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.2.0
executing==1.2.0
filelock==3.12.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
virtualenv==20.24.6
wcwidth==0.2.6
